home-assistant-cdf-extractor/pyproject.toml

[tool.poetry]
name = "home_assistant_cdf_extractor"
version = "1.0.0"
description = "Extract data from Home Assistant"
authors = ["Tobias Slettemoen Kongsvik <tobias.kongsvik@cognite.com>"]

packages = [
    {include = "home_assistant_cdf_extractor"}
]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["arrow", "cognite", "dateutil", "requests"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
cognite-sdk-core = "^2.50.1"
cognite-extractor-utils = "^2.2.0"
cognite-extractor-utils-rest = { git = "https://github.com/cognitedata/python-extractor-utils-rest", branch = "tobias/json-root-list"}
arrow = "^1.2.2"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.7"
macholib = {version = "^1.14", platform = "darwin"}             # Used by pyinstaller pn Mac OS
pywin32-ctypes = {version = "^0.2.0", platform = "win32"}       # Used by pyinstaller on Windows
pefile = "^2019.4.18"                                           # Used by pyinstaller on Windows
mypy = "^0.961"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"

[tool.poetry.scripts]
home_assistant_cdf_extractor = "home_assistant_cdf_extractor.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
